Flatten the select FIO address styles
